fix(test): pr comments fix
